+++ PlayCover/Utils/DataExtensions.swift
@@ -1,0 +1,29 @@
+//
+//  DataExtensions.swift
+//  PlayCover
+//
+
+import Foundation
+
+extension String {
+    init(data: Data, offset: Int, commandSize: Int, loadCommandString: lc_str) {
+        let loadCommandStringOffset = Int(loadCommandString.offset)
+        let stringOffset = offset + loadCommandStringOffset
+        let length = commandSize - loadCommandStringOffset
+        self = String(data: data[stringOffset..<(stringOffset + length)],
+                      encoding: .utf8)!.trimmingCharacters(in: .controlCharacters)
+    }
+}
+
+extension Data {
+    func extract<T>(_ type: T.Type, offset: Int = 0) -> T {
+        let data = self[offset..<offset + MemoryLayout<T>.size]
+        return data.withUnsafeBytes { dataBytes in
+            dataBytes.baseAddress!
+                .assumingMemoryBound(to: UInt8.self)
+                .withMemoryRebound(to: T.self, capacity: 1) { (pointer) -> T in
+                return pointer.pointee
+            }
+        }
+    }
+}
