+++ PlayCover/Utils/URLExtensions.swift
@@ -1,5 +1,5 @@
 //
-//  Utils.swift
+//  URLExtensions.swift
 //  PlayCover
 //
 
@@ -27,7 +27,6 @@ extension String {
 
 extension URL {
     func subDirectories() throws -> [URL] {
-        // @available(macOS 10.11, iOS 9.0, *)
         guard hasDirectoryPath else { return [] }
         return try FileManager.default
             .contentsOfDirectory(at: self, includingPropertiesForKeys: nil, options: [.skipsHiddenFiles])
@@ -105,5 +104,9 @@ extension URL {
         }
 
         return self.appendingPathComponent(newPathComponent)
+    }
+
+    func setBinaryPosixPermissions(_ permissions: Int) throws {
+        try FileManager.default.setAttributes([.posixPermissions: permissions], ofItemAtPath: path)
     }
 }
